requirements.txt
pandas
azure-storage-blob
python-dotenv